requirements.txt
fastapi==0.112.0
uvicorn==0.30.4
pillow==10.4.0
mtcnn==0.1.1
numpy<2
torch==2.4.0
torchvision==0.19.0
cmake==3.30.1
tensorflow==2.14.0
stylegan2-pytorch==1.8.10